Make all embedded fields on runner private
